Refactor: move types to their own files from services
