+++ src/app/haku/[oid]/hakukohde/[hakukohde]/perustiedot/valintatapajonot-table.tsx
@@ -6,7 +6,7 @@ import ListTable, {
 } from '@/app/components/table/list-table';
 import { isToisenAsteenYhteisHaku } from '@/app/lib/kouta';
 import { Haku } from '@/app/lib/kouta-types';
-import { ValintatapajonoTulos } from '@/app/lib/valinta-tulos-service';
+import { ValintatapajonoTulos } from '@/app/lib/sijoittelu-types';
 
 export const ValintatapajonotTable = ({
   valintatapajonoTulokset,
