+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valinnan-hallinta/hallinta-table-row.tsx
@@ -1,9 +1,6 @@
 'use client';
 
-import {
-  Valinnanvaihe,
-  isLaskentaUsedForValinnanvaihe,
-} from '@/app/lib/valintaperusteet';
+import { isLaskentaUsedForValinnanvaihe } from '@/app/lib/valintaperusteet';
 import { Box, CircularProgress, TableCell, TableRow } from '@mui/material';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { Button } from '@opetushallitus/oph-design-system';
@@ -22,6 +19,7 @@ import { Haku, Hakukohde } from '@/app/lib/kouta-types';
 import { sijoitellaankoHaunHakukohteetLaskennanYhteydessa } from '@/app/lib/kouta';
 import { HaunAsetukset } from '@/app/lib/ohjausparametrit';
 import { useToaster } from '@/app/hooks/useToaster';
+import { Valinnanvaihe } from '@/app/lib/valintaperusteet-types';
 
 type HallintaTableRowParams = {
   haku: Haku;
