+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintalaskennan-tulos/valintatapajono-content.tsx
@@ -1,6 +1,6 @@
 'use client';
 import { Box } from '@mui/material';
-import { LaskettuValinnanVaihe } from '@/app/lib/valintalaskenta-service';
+import { LaskettuValinnanVaihe } from '@/app/lib/laskenta-types';
 import React from 'react';
 import { ValintatapajonoAccordion } from './valintatapajono-accordion';
 import { useJonosijatSearch } from '@/app/hooks/useJonosijatSearch';
