+++ src/app/hooks/useHakeneetSearch.ts
@@ -10,7 +10,8 @@ import {
   HAKU_SEARCH_PHRASE_DEBOUNCE_DELAY,
 } from '@/app/lib/constants';
 import { useTranslations } from './useTranslations';
-import { Hakemus, getHakemukset } from '../lib/ataru';
+import { getHakemukset } from '../lib/ataru';
+import { Hakemus } from '../lib/ataru-types';
 
 const DEFAULT_NUQS_OPTIONS = {
   history: 'push',
