+++ src/app/lib/valinta-tulos-service.ts
@@ -2,19 +2,12 @@
 
 import { configuration } from './configuration';
 import { client } from './http-client';
-
-export type ValintatapajonoTulos = {
-  nimi: string;
-  oid: string;
-  sijoittelunAloituspaikat: string;
-  hyvaksytty: number;
-  ehdollisestiHyvaksytty: number;
-  harkinnanvaraisestiHyvaksytty: number;
-  varasijoilla: number;
-  vastaanottaneet: number;
-  paikanPeruneet: number;
-  pisteraja: number;
-};
+import {
+  SijoitteluajonTulokset,
+  SijoittelunHakemus,
+  SijoittelunTila,
+  ValintatapajonoTulos,
+} from './sijoittelu-types';
 
 export const getSijoittelunTulokset = async (
   hakuOid: string,
@@ -53,50 +46,6 @@ export const getSijoittelunTulokset = async (
     },
   );
   return jsonTulokset;
-};
-
-export enum SijoittelunTila {
-  HYVAKSYTTY = 'HYVAKSYTTY',
-  VARASIJALTA_HYVAKSYTTY = 'VARASIJALTA_HYVAKSYTTY',
-  HARKINNANVARAISESTI_HYVAKSYTTY = 'HARKINNANVARAISESTI_HYVAKSYTTY',
-  VARALLA = 'VARALLA',
-  HYLATTY = 'HYLATTY',
-  PERUUNTUNUT = 'PERUUNTUNUT',
-  PERUNUT = 'PERUNUT',
-  PERUUTETTU = 'PERUUTETTU',
-}
-
-export const SijoittelunTilaOrdinals: Record<string, number> = Object.keys(
-  SijoittelunTila,
-)
-  .map((key, index) => ({ [key]: index }))
-  .reduce((a, b) => ({ ...a, ...b }));
-
-export type SijoittelunHakemus = {
-  hakijaOid: string;
-  hakemusOid: string;
-  pisteet: number;
-  tila: SijoittelunTila;
-  valintatapajonoOid: string;
-  hyvaksyttyHakijaryhmista: string[];
-  varasijanNumero: number;
-};
-
-export type SijoitteluajonValintatapajono = {
-  oid: string;
-  nimi: string;
-  hakemukset: SijoittelunHakemus[];
-  prioriteetti: number;
-};
-
-export type SijoittelunHakijaryhmat = {
-  oid: string;
-  kiintio: number;
-};
-
-export type SijoitteluajonTulokset = {
-  valintatapajonot: SijoitteluajonValintatapajono[];
-  hakijaryhmat: SijoittelunHakijaryhmat[];
 };
 
 export const getLatestSijoitteluAjonTulokset = async (
